update: move model download screen to main screen
