+++ app/src/main/res/values/strings.xml
@@ -51,7 +51,7 @@
     <string name="download_model_action">Download model</string>
 
 
-    <!--Alerts-->
+    <!--Alerts / Message-->
     <string name="alert_space_not_allowed">Spaces are not allowed</string>
     <string name="alert_scan_refresh_description">Manually update the index with new media (auto-updates daily).</string>
     <string name="alert_scan_rebuild_description">Delete and re-create the entire index.</string>
@@ -62,6 +62,7 @@
     <string name="scan_in_progress_content">This may take a moment. You can continue using the app once the scan completes.</string>
     <string name="media_type_title">Media type</string>
     <string name="alert_initial_scan_required">To use this feature, you need to complete the initial scan of your library first.</string>
+    <string name="download_model_message">A new model is required to use the app. Download it now to continue.</string>
 
     <!-- Search screen -->
     <string name="search_storage_permissions">Storage permission required to search.</string>
@@ -81,7 +82,6 @@
     <string name="concepts_step_generate_summaries">Generate descriptions for images in your selected collections.</string>
     <string name="concepts_required_api_key_note">A SmartScan API key is required to generate descriptions.</string>
     <string name="concepts_storage_permissions">Storage permission required to create concepts.</string>
-    <string name="concepts_download_model">A new model is required to use Concepts. Download it now to enable this feature. You can manage this model anytime from Settings.</string>
     <string name="concepts_delete_alert_title">Delete concepts</string>
     <string name="concepts_delete_alert_description">Delete %1$d %2$s?</string>
     <plurals name="concept_count">
